Rename gen_expl dir to explorer_common. Rename a bunch of other grab-bag directories and modules that weren't in line with our naming conventions
