+++ client/src/IgocExplorer/IgocExplorer.scss
