+++ client/src/IgocExplorer/hierarchies.js
@@ -1,5 +1,5 @@
 import { businessConstants } from '../models/businessConstants.js';
-import { convert_d3_hierarchy_to_explorer_hierarchy } from '../gen_expl/hierarchy_tools.js';
+import { convert_d3_hierarchy_to_explorer_hierarchy } from '../explorer_common/hierarchy_tools.js';
 import { Subject } from '../models/subject.js';
 import { igoc_tmf as text_maker } from './igoc_explorer_text.js';
