+++ client/src/IgocExplorer/igoc_explorer_text.js
@@ -1,4 +1,4 @@
-import igoc_explorer_bundle from './igoc_explorer.yaml';
+import igoc_explorer_bundle from './IgocExplorer.yaml';
 import { create_text_maker_component } from '../components/index.js';
 
 export const { text_maker: igoc_tmf, TM } = create_text_maker_component(igoc_explorer_bundle);
