+++ client/src/InfoLab/InfoLab.scss
