+++ client/src/explorer_common/explorer-styles.scss
