+++ client/src/explorer_common/hierarchy_tools.js
