+++ client/src/explorer_common/resource_explorer_common.js
