+++ client/src/explorer_common/resource_scheme.js
@@ -1,6 +1,6 @@
 import { createSelector } from 'reselect';
 import { infograph_href_template } from '../link_utils.js';
-import { provide_sort_func_selector } from './resource-explorer-common.js';
+import { provide_sort_func_selector } from './resource_explorer_common.js';
 import { shallowEqualObjectsOverKeys, sanitized_dangerous_inner_html } from '../general_utils.js';
 import { HeightClipper } from '../components/index.js';
 import { get_resources_for_subject } from './resource_utils.js';
