+++ client/src/explorer_common/resource_utils.js
