+++ client/src/explorer_common/result_displays.js
