+++ client/src/explorer_common/result_hierarchies.js
