+++ client/src/explorer_common/results_scheme.js
