+++ client/src/explorer_common/search_tools.js
