+++ client/src/explorer_common/state_and_memoizing.js
