+++ client/src/infographic/Infographic.scss
