+++ client/src/infographic/Infographic.yaml
