+++ client/src/panels/panel_declarations/misc/resource_structure.js
@@ -1,4 +1,4 @@
-import '../../../gen_expl/explorer-styles.scss';
+import '../../../explorer_common/explorer-styles.scss';
 import text from './resource_structure.yaml';
 
 import { combineReducers, createStore } from 'redux';
@@ -7,8 +7,8 @@ import { Fragment } from 'react';
 import { Explorer } from '../../../components/ExplorerComponents.js';
 
 import { infograph_href_template } from '../../../link_utils.js';
-import { get_root } from '../../../gen_expl/hierarchy_tools.js';
-import { get_col_defs } from '../../../gen_expl/resource-explorer-common.js';
+import { get_root } from '../../../explorer_common/hierarchy_tools.js';
+import { get_col_defs } from '../../../explorer_common/resource_explorer_common.js';
 
 import {
   declare_panel,
@@ -20,7 +20,7 @@ import {
 import {
   create_rooted_resource_scheme,
   get_initial_resource_state,
-} from '../../../gen_expl/rooted_resource_scheme.js';
+} from '../../../explorer_common/rooted_resource_scheme.js';
 
 import {
   get_memoized_funcs,
@@ -28,7 +28,7 @@ import {
   root_reducer,
   map_state_to_root_props_from_memoized_funcs,
   map_dispatch_to_root_props,
-} from '../../../gen_expl/state_and_memoizing';
+} from '../../../explorer_common/state_and_memoizing';
 
 
 const { text_maker, TM } = create_text_maker_component(text);
